AI_MODEL/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
ultralytics
